fix: dismiss by swipe
